Move banner components to components folder
This stops error messages due to name clashing
in esm imports
